+++ HISTORY.rst
@@ -84,8 +84,11 @@ Changes
 * Smoothing, simplification and regularization of boundaries implemented
 * Colortable for legend is persistant trough indexing. Classify method
   could be used to change it
-* Default color table is rainbow for unique and viridis for continuous classes
+* Default color table is seaborn muted for unique classification
+  and matplotlib viridis for continuous classes
 
 0.5 (XX YYY 2017)
 -----------------
- * rose method groupped according to classification
+ * rose plot groupped according to classification
+ * get_class, class_iter methods added to Grains and Boundaries
+ * seaborn added to requirements